# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invest Agents Web UI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/portfolio/actual/transactions": {
      "get": {
        "summary": "Get Actual Transactions",
        "operationId": "get_actual_transactions_api_portfolio_actual_transactions_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "portfolio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/actual/transactions/{tx_id}/note": {
      "patch": {
        "summary": "Update Transaction Note Endpoint",
        "operationId": "update_transaction_note_endpoint_api_portfolio_actual_transactions__tx_id__note_patch",
        "parameters": [
          {
            "name": "tx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tx Id"
            }
          },
          {
            "name": "portfolio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionNoteRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionItemDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/actual/transactions/{tx_id}": {
      "put": {
        "summary": "Edit Transaction Endpoint",
        "operationId": "edit_transaction_endpoint_api_portfolio_actual_transactions__tx_id__put",
        "parameters": [
          {
            "name": "tx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tx Id"
            }
          },
          {
            "name": "portfolio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTransactionRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActualPortfolioStateDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Transaction Endpoint",
        "operationId": "delete_transaction_endpoint_api_portfolio_actual_transactions__tx_id__delete",
        "parameters": [
          {
            "name": "tx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tx Id"
            }
          },
          {
            "name": "adjust_cash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Adjust Cash"
            }
          },
          {
            "name": "portfolio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActualPortfolioStateDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/actual/transactions/{tx_id}/void": {
      "post": {
        "summary": "Void Transaction Endpoint",
        "operationId": "void_transaction_endpoint_api_portfolio_actual_transactions__tx_id__void_post",
        "parameters": [
          {
            "name": "tx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tx Id"
            }
          },
          {
            "name": "portfolio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActualPortfolioStateDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualHoldingDTO": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "asset_type": {
            "type": "string",
            "title": "Asset Type"
          },
          "units": {
            "type": "number",
            "title": "Units"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "bucket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket Id"
          },
          "avg_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cost Usd"
          },
          "avg_cost_thb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cost Thb"
          },
          "current_price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price Usd"
          },
          "current_price_thb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price Thb"
          },
          "fx_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fx Rate"
          },
          "market_value_thb": {
            "type": "number",
            "title": "Market Value Thb",
            "default": 0.0
          },
          "unrealized_pnl_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Pnl Percent"
          },
          "unrealized_pnl_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Pnl Value"
          },
          "market_cap_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Tier"
          },
          "yield_on_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yield On Cost"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio"
          },
          "eps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eps"
          },
          "payout_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payout Ratio"
          },
          "market_cap_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Value"
          },
          "dividend_per_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Per Share"
          },
          "dividend_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield"
          },
          "accumulated_dividend_thb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accumulated Dividend Thb"
          },
          "accumulated_dividend_native": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accumulated Dividend Native"
          },
          "upcoming_dividend_thb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upcoming Dividend Thb"
          },
          "upcoming_dividend_native": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upcoming Dividend Native"
          },
          "dividend_rounds": {
            "items": {
              "$ref": "#/components/schemas/DividendRoundDTO"
            },
            "type": "array",
            "title": "Dividend Rounds"
          },
          "dividend_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "synced",
                  "manual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Source"
          },
          "fundamentals_updated_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fundamentals Updated At"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "asset_type",
          "units"
        ],
        "title": "ActualHoldingDTO",
        "description": "12-Column Holding schema matching Frontend Holdings Table & Backend Holding Model"
      },
      "ActualPortfolioStateDTO": {
        "properties": {
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "fx_rates": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Fx Rates",
            "default": {}
          },
          "summary": {
            "$ref": "#/components/schemas/ActualSummaryDTO"
          },
          "allocation_targets": {
            "items": {
              "$ref": "#/components/schemas/AllocationTargetDTO"
            },
            "type": "array",
            "title": "Allocation Targets",
            "default": []
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/ActualHoldingDTO"
            },
            "type": "array",
            "title": "Holdings",
            "default": []
          },
          "price_refresh_info": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Refresh Info"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "ActualPortfolioStateDTO"
      },
      "ActualSummaryDTO": {
        "properties": {
          "total_value_thb": {
            "type": "number",
            "title": "Total Value Thb",
            "default": 0.0
          },
          "total_cost_basis_thb": {
            "type": "number",
            "title": "Total Cost Basis Thb",
            "default": 0.0
          },
          "total_unrealized_profit": {
            "type": "number",
            "title": "Total Unrealized Profit",
            "default": 0.0
          },
          "total_realized_profit_ytd": {
            "type": "number",
            "title": "Total Realized Profit Ytd",
            "default": 0.0
          },
          "passive_income_ytd": {
            "type": "number",
            "title": "Passive Income Ytd",
            "default": 0.0
          },
          "total_accumulated_dividend": {
            "type": "number",
            "title": "Total Accumulated Dividend",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "ActualSummaryDTO"
      },
...
      "AllocationTargetDTO": {
        "properties": {
          "bucket_id": {
            "type": "string",
            "title": "Bucket Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Target Percent"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "bucket_id",
          "name",
          "target_percent"
        ],
        "title": "AllocationTargetDTO"
      },
...
      "DividendRoundDTO": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "ex_date": {
            "type": "string",
            "title": "Ex Date"
          },
          "pay_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pay Date"
          },
          "dps": {
            "type": "number",
            "title": "Dps"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "units_held": {
            "type": "number",
            "title": "Units Held"
          },
          "status": {
            "type": "string",
            "enum": [
              "received",
              "upcoming"
            ],
            "title": "Status",
            "default": "received"
          },
          "gross_native": {
            "type": "number",
            "title": "Gross Native",
            "default": 0.0
          },
          "net_native": {
            "type": "number",
            "title": "Net Native",
            "default": 0.0
          },
          "gross_thb": {
            "type": "number",
            "title": "Gross Thb",
            "default": 0.0
          },
          "tax_rate": {
            "type": "number",
            "title": "Tax Rate",
            "default": 0.0
          },
          "net_thb": {
            "type": "number",
            "title": "Net Thb",
            "default": 0.0
          },
          "fx_rate": {
            "type": "number",
            "title": "Fx Rate",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "ex_date",
          "dps",
          "currency",
          "units_held"
        ],
        "title": "DividendRoundDTO"
      },
...
      "EditTransactionRequestDTO": {
        "properties": {
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "units": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "fx_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fx Rate"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "adjust_cash": {
            "type": "boolean",
            "title": "Adjust Cash",
            "default": true
          }
        },
        "type": "object",
        "title": "EditTransactionRequestDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionItemDTO": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "units": {
            "type": "number",
            "title": "Units"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "fx_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fx Rate"
          },
          "cost_thb": {
            "type": "number",
            "title": "Cost Thb",
            "default": 0.0
          },
          "realized_pnl_thb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Pnl Thb"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "gross_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Amount"
          },
          "commission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission"
          },
          "vat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vat"
          },
          "other_fees": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Fees"
          },
          "net_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Amount"
          },
          "fee_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Currency"
          },
          "confirmation_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation No"
          },
          "settlement_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settlement Date"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "cash_adjusted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Adjusted"
          },
          "related_transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "timestamp",
          "symbol",
          "action",
          "units",
          "price",
          "currency"
        ],
        "title": "TransactionItemDTO"
      },
      "TransactionListResponseDTO": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "title": "Portfolio Id"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionItemDTO"
            },
            "type": "array",
            "title": "Transactions",
            "default": []
          },
          "summary": {
            "$ref": "#/components/schemas/TransactionSummaryDTO"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id"
        ],
        "title": "TransactionListResponseDTO"
      },
      "TransactionSummaryDTO": {
        "properties": {
          "total_buy_count": {
            "type": "integer",
            "title": "Total Buy Count",
            "default": 0
          },
          "total_sell_count": {
            "type": "integer",
            "title": "Total Sell Count",
            "default": 0
          },
          "total_buy_thb": {
            "type": "number",
            "title": "Total Buy Thb",
            "default": 0.0
          },
          "total_sell_thb": {
            "type": "number",
            "title": "Total Sell Thb",
            "default": 0.0
          },
          "total_realized_pnl_thb": {
            "type": "number",
            "title": "Total Realized Pnl Thb",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "TransactionSummaryDTO"
      },
...
      "UpdateTransactionNoteRequestDTO": {
        "properties": {
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "title": "UpdateTransactionNoteRequestDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}